youtube_transcript_pipeline/src/yt_transcripts/config.py
from __future__ import annotations

import logging
import os

from dotenv import load_dotenv


load_dotenv()


def configure_logging() -> None:
    logging.basicConfig(
        level=os.getenv("LOG_LEVEL", "INFO").upper(),
        format="%(asctime)s | %(levelname)s | %(name)s | %(message)s",
    )
