youtube_transcript_pipeline/requirements.txt
python-dotenv
youtube-transcript-api
yt-dlp
google-api-python-client
openai-whisper
